better docs and refactor for name
