+++ src/main/scala/ai/eto/rikai/sql/spark/RikaiSparkSessionExtensions.scala
@@ -17,7 +17,7 @@ package ai.eto.rikai.sql.spark
 import ai.eto.rikai.sql.spark.expressions.Predict
 import ai.eto.rikai.sql.spark.parser.{RikaiExtSqlParser, RikaiSparkSQLParser}
 import org.apache.spark.sql.SparkSessionExtensions
-import org.apache.spark.sql.rikai.RikaiTypeRegisters
+import org.apache.spark.sql.rikai.UDTCollection
 
 /**
   * Rikai SparkSession extensions to enable Spark SQL ML.
@@ -27,7 +27,7 @@ class RikaiSparkSessionExtensions extends (SparkSessionExtensions => Unit) {
   override def apply(extensions: SparkSessionExtensions): Unit = {
 
     extensions.injectParser((session, parser) => {
-      RikaiTypeRegisters.loadUDTs(session)
+      UDTCollection.loadUDTs(session)
 
       new RikaiExtSqlParser(
         session,
